Move build task registrations out of command and into task service
Fixes https://github.com/hydephp/develop/issues/1175
